Quality-of-Life Changes
Less scrolling is required to change the pen size.
The canvas can be zoomed in/out of.
Panning can alternatively be done with the middle mouse button.
Added controls instructions on the top bar.
Added client info to the footer.
A list of online players is displayed on the right.